tests/wpt/tests/dom/nodes/insertion-removing-steps/Node-appendChild-script-in-script.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting a script and some code in an empty script</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<body>
<script id="s1"></script>
<script>
const happened = [];
test(() => {
  const s1 = document.getElementById("s1");
  const s2 = document.createElement("script");

  // This script, which is ultimately a *child* of the
  // already-connected-but-empty `s1` script, runs first. This is because when
  // the DocumentFragment `df` containing this script (`s2`) is appended to
  // `s1`, we do not immediately execute `s1`, because of this condition: [1].
  // It specifies that when a "node or document fragment is inserted into the
  // script", we only run the "prepare the script algorithm" for that script
  // "after any script elements inserted at that time".
  //
  // So upon insertion of `s2`, its insertion conditions are met and it is
  // prepared and run first. During its execution, it mutates `s1`, adding more
  // text to it. This condition still does not trigger the immediate execution
  // of `s1`, because `s2`'s insertion is not complete yet.
  //
  // Once `s2` is finished being inserted, the condition to prepare and execute
  // `s1` is met, and it is processed accordingly, which includes the execution
  // of the text that `s2` added to `s1`.
  //
  // [1]: https://html.spec.whatwg.org/C#script-processing-model:prepare-the-script-element-4
  s2.textContent = `
    happened.push("s2");
    s1.appendChild(new Text("happened.push('s1ran');"));
    happened.push("s2ran");
`;

  const df = document.createDocumentFragment();
  df.appendChild(new Text(`happened.push("s1");`));
  df.appendChild(s2);

  assert_array_equals(happened, []);
  s1.appendChild(df);
  assert_array_equals(happened, ["s2", "s2ran", "s1", "s1ran"]);
}, "An outer script does not execute until its inner `<script>` children are " +
   "finished being inserted into it. The outer script's execution considers " +
   "any script text that inner scripts added to the outer one.");
</script>
